% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsPack.R
\name{generateDummyVariables}
\alias{generateDummyVariables}
\title{Conversion factors in dummy}
\usage{
generateDummyVariables(NaNData)
}
\arguments{
\item{NaNData}{data.frame of one ore several factors columns}
}
\value{
a data.frame in which each variable are represented as dummies.
}
\description{
This function returns a data.frame of dummy corresponding to a data.frame of factors
representing a categorical variable.
}
