% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsPack.R
\name{get_eigenvec}
\alias{get_eigenvec}
\title{Eigenvectors calculations}
\usage{
get_eigenvec(tree, variance_fraction = 0.8, numEigen = NULL)
}
\arguments{
\item{tree}{phylogenetic tree of class "phylo"}

\item{variance_fraction}{total amount of minimum variance to be represented by the eigenvectors which correspond
to the phylogenetic inertia}
}
\value{
a data.frame in which each colum represent an eigenvector
}
\description{
This function calculates the eigenvectors of a phylo object
}
