% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeKNN.R
\name{imputeKNN}
\alias{imputeKNN}
\title{non-parametric missing value imputation for mixed-type data by kNN(k-nearest neighbor algorithm)}
\usage{
imputeKNN(missingData, k, numFun, catFun, variance_fraction = 0.8, tree = NULL)
}
\arguments{
\item{missingData}{data.frame of 1 or more columns containing NAs}

\item{k}{integer, number of nearest neighbours used}

\item{numFun}{numFun: function for aggregating the kNN in the case of a numerical variable}

\item{catFun}{catFun: function for aggregating the kNN in the case of a categorical variable}

\item{variance_fraction}{total amount of minimum variance to be represented by the eigenvectors which correspond
to the phylogenetic inertia}

\item{tree}{phylo object}

\item{hint}{dataframe already imputed by a comprative methods (Rphylopars for continuous and corHMM for discrete)}
}
\value{
a data.frame of 1 or more numeric columns with the NAs replaced by values + parameters used for the imputation
}
\description{
This function imputes missing data for continuous and discrete traits applying the kNN approach
}
