% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeMICE.R
\name{imputeMICE}
\alias{imputeMICE}
\title{Imputation of missing data for continuous and discrete traits by MICE approach}
\usage{
imputeMICE(missingData, nbrMI, method, tree)
}
\arguments{
\item{missingData}{data.frame of 1 or more columns containing NAs}

\item{nbrMI}{integer, mentioning the total number of imputations}

\item{variance_fraction}{total amount of minimum variance to be represented by the eigenvectors which correspond
to the phylogenetic inertia}

\item{tree}{phylo object}

\item{hint}{dataframe already imputed by a comprative methods (Rphylopars for continuous and corHMM for discrete)}
}
\value{
a data.frame of 1 or more numeric columns with the NAs replaced by values + parameters used for the
imputation
}
\description{
This function imputes missing data for continuous and discrete traits applying the MICE approach
using the predictive mean matching method for continuous traits, the polytomous regression for unordered
categorical traits having >2 states, if binary the algo uses a logistic regression and in case of ordered traits
with >2 states, the method uses proportional odds model.
}
