% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeMissForest.R
\name{imputeMissForest}
\alias{imputeMissForest}
\title{non-parametric missing value imputation for mixed-type data by missForest(randomForest)}
\usage{
imputeMissForest(missingData, variance_fraction = 0.8, maxiter = 10, ntree = 100, 
mtry = sqrt(ncol(missingData)), tree = NULL)
}
\arguments{
\item{missingData}{data.frame of 1 or more columns containing NAs}

\item{variance_fraction}{total amount of minimum variance to be represented by the eigenvectors which correspond
to the}

\item{maxiter}{maximum number of iterations to be performed given the stopping criterion is not met beforehand.}

\item{ntree}{number of trees to grow in each forest.}

\item{mtry}{number of variables randomly sampled at each split. By default it's the square root of the number of
variables}

\item{tree}{phylo object
phylogenetic inertia}

\item{hint}{dataframe already imputed by a comprative methods (Rphylopars for continuous and corHMM for discrete)}
}
\value{
a data.frame of 1 or more numeric columns with the NAs replaced by values. + parameters used for the imputation
}
\description{
This function imputes missing data for continuous and discrete traits applying the missForest
approach
}
