% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputePIDiscreteTraits.R
\name{imputeOneDiscreteTrait}
\alias{imputeOneDiscreteTrait}
\title{Imputation of missing data for one discrete trait}
\usage{
imputeOneDiscreteTrait(trait, Data)
}
\arguments{
\item{missingData}{data.frame of 1 factor column containing NAs}

\item{tree}{phylo}
}
\value{
a data.frame of 1 factor column with the NAs replaced by values
}
\description{
This function imputes missing data for discrete traits using the R package phytools. The first step
is to run the function make.simmap fits a continuous-time reversible Markov model for the evolution of one trait
and then simulates stochastic character histories using that model and the tip states on the tree.
for three models(ER, SYM and ARD) and select the one having the smallest AIC.
}
