% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputePIDiscreteTraits.R
\name{imputePIDiscreteTraits}
\alias{imputePIDiscreteTraits}
\title{Imputation of missing data for discrete traits, mandatory phylogenetic information}
\usage{
PIDiscreteTrait(trait)
}
\arguments{
\item{missingData}{data.frame of 1 or more factor columns containing NAs}

\item{tree}{phylo object}
}
\value{
a data.frame of 1 or more factor columns with the NAs replaced by values, parameters used for the
imputation, matrix of probabilities of each states
}
\description{
This function apply the function imputeOneDiscretTrait on several columns at one time.
}
